requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
google-cloud-storage
pyyaml
mlflow
dagshub 